Fixed unit tests (and moved them to `chain.rs` where actual method is now defined)
