+++ node/src/main.rs
@@ -648,100 +648,3 @@ fn start_firehose_block_ingestor<C, M>(
             }
         });
 }
-#[cfg(test)]
-mod test {
-    use graph::log::logger;
-    use graph::prelude::tokio;
-    use graph::prometheus::Registry;
-    use graph_chain_ethereum::NodeCapabilities;
-    use graph_core::MetricsRegistry;
-    use std::sync::Arc;
-
-    #[tokio::test]
-    async fn correctly_parse_ethereum_networks() {
-        let logger = logger(true);
-
-        let network_args = vec![
-            "mainnet:traces:http://localhost:8545/".to_string(),
-            "goerli:archive:http://localhost:8546/".to_string(),
-        ];
-
-        let opt = Opt {
-            postgres_url: Some("not needed".to_string()),
-            config: None,
-            store_connection_pool_size: 5,
-            postgres_secondary_hosts: vec![],
-            postgres_host_weights: vec![],
-            disable_block_ingestor: true,
-            node_id: "default".to_string(),
-            ethereum_rpc: network_args,
-            ethereum_ws: vec![],
-            ethereum_ipc: vec![],
-            unsafe_config: false,
-        };
-
-        let config = Config::load(&logger, &opt).expect("can create config");
-        let prometheus_registry = Arc::new(Registry::new());
-        let metrics_registry = Arc::new(MetricsRegistry::new(
-            logger.clone(),
-            prometheus_registry.clone(),
-        ));
-
-        let ethereum_networks = create_ethereum_networks(logger, metrics_registry, config.clone())
-            .await
-            .expect("Correctly parse Ethereum network args");
-        let mut network_names = ethereum_networks.networks.keys().collect::<Vec<&String>>();
-        network_names.sort();
-
-        let traces = NodeCapabilities {
-            archive: false,
-            traces: true,
-        };
-        let archive = NodeCapabilities {
-            archive: true,
-            traces: false,
-        };
-        let has_mainnet_with_traces = ethereum_networks
-            .adapter_with_capabilities("mainnet".to_string(), &traces)
-            .is_ok();
-        let has_goerli_with_archive = ethereum_networks
-            .adapter_with_capabilities("goerli".to_string(), &archive)
-            .is_ok();
-        let has_mainnet_with_archive = ethereum_networks
-            .adapter_with_capabilities("mainnet".to_string(), &archive)
-            .is_ok();
-        let has_goerli_with_traces = ethereum_networks
-            .adapter_with_capabilities("goerli".to_string(), &traces)
-            .is_ok();
-
-        assert_eq!(has_mainnet_with_traces, true);
-        assert_eq!(has_goerli_with_archive, true);
-        assert_eq!(has_mainnet_with_archive, false);
-        assert_eq!(has_goerli_with_traces, false);
-
-        let goerli_capability = ethereum_networks
-            .networks
-            .get("goerli")
-            .unwrap()
-            .adapters
-            .iter()
-            .next()
-            .unwrap()
-            .capabilities;
-        let mainnet_capability = ethereum_networks
-            .networks
-            .get("mainnet")
-            .unwrap()
-            .adapters
-            .iter()
-            .next()
-            .unwrap()
-            .capabilities;
-        assert_eq!(
-            network_names,
-            vec![&"goerli".to_string(), &"mainnet".to_string()]
-        );
-        assert_eq!(goerli_capability, archive);
-        assert_eq!(mainnet_capability, traces);
-    }
-}
